scope dfc routes with v0 api version
