+++ engines/dfc_provider/app/serializers/dfc_provider/supplied_product_serializer.rb
@@ -17,7 +17,7 @@ module DfcProvider
     has_many :physical_characteristics, key: 'dfc:physicalCharacterisctics'
 
     def id
-      dfc_provider_routes.api_dfc_provider_enterprise_supplied_product_url(
+      dfc_provider_routes.api_v0_dfc_provider_enterprise_supplied_product_url(
         enterprise_id: object.product.supplier_id,
         id: object.id,
         host: host
