+++ engines/dfc_provider/config/routes.rb
@@ -7,12 +7,14 @@ DfcProvider::Engine.routes.draw do
     end
 
     namespace :api do
-      scope :dfc_provider, as: :dfc_provider, path: '/dfc_provider' do
-        resources :enterprises, only: [:show] do
-          resources :catalog_items, only: [:index, :show]
-          resources :supplied_products, only: [:show]
+      namespace :v0 do
+        scope :dfc_provider, as: :dfc_provider, path: '/dfc_provider' do
+          resources :enterprises, only: [:show] do
+            resources :catalog_items, only: [:index, :show]
+            resources :supplied_products, only: [:show]
+          end
+          resources :persons, only: [:show]
         end
-        resources :persons, only: [:show]
       end
     end
   end
